requirements.txt
discord.py>=2.4.0
anthropic>=0.37.1
python-dotenv>=1.0.1
aiohttp>=3.9.1
pandas>=2.2.0
pytest>=8.0.0
python-dateutil>=2.9.0
anyio>=4.0.0
httpx>=0.27.0